+++ NEWS.md
@@ -1,3 +1,8 @@
+# PAnEn 4.2.5
+
+- Added `RAnEn::readNc2`
+- Added `recursive` to `RAnEn::print.list`
+
 # PAnEn 4.2.4
 
 - Moved 'RAnEn::schaakeShuffle' to `RAnEnExtra::schaakeShuffle`